% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{My k-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{input data frame.}

\item{cl}{true class value of your training data.}

\item{k_nn}{integer representing the number of neighbors.}

\item{k_cv}{integer representing the number of folds.}
}
\value{
a \code{list} with:

\code{class} with elements:

\code{cv_err} the numeric test statistic.
}
\description{
This function can predict an unknown class using other known classes.
}
\examples{
my_knn_cv(penguins_train, penguins_knn$species, 1, 5)
my_knn_cv(penguins_train, penguins_knn$species, 5, 5)


}
\keyword{prediction}
