% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{My Linear Model Function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a \code{formula} class object, similar to \code{lm()}.}

\item{data}{input data frame.}
}
\value{
a \code{table} similar to the coefficent table from \code{summary()}
with rows for each coefficient (including the \code{(Intercept)}!)
and columns for the \code{Estimate}, \code{Std. Error}, \code{t value},
and \code{Pr(>|t|)}. There should be row and column names.
}
\description{
This function fits a linear model in R.
}
\examples{
my_lm(eruptions ~ waiting, faithful)
summary(lm(eruptions ~ waiting, data=faithful))


}
\keyword{inference}
