% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{My t-test Function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{a numeric vector of data.}

\item{alternative}{a character string specifying the alternative hypothesis.
This should only accept \code{"two.sided"}, \code{"less"},
or \code{"greater"}. Otherwise, your function should
throw an informative error.}

\item{mu}{a number indicating the null hypothesis value of the mean.}
}
\value{
a \code{list} with elements:

\code{test_stat}: the numeric test statistic.

\code{df}: the degrees of freedom.

\code{alternative}: the value of the parameter \code{alternative}

\code{p_val}: the numeric p-value.
}
\description{
This function performs a one sample t-test in R.
}
\examples{
data <- c(1, 3, 6, 9, 10, 28, 37, 46, 55, 83, 27, 39, 42, 63, 100)
my_t.test(data, "two.sided", 60)
my_t.test(data, "less", 60)
my_t.test(data, "greater", 60)


}
\keyword{inference}
